+++ History.md
@@ -1,3 +1,8 @@
+[3.2.11 / 2015-xx-xx](https://github.com/jakubpawlowicz/clean-css/compare/v3.2.10...3.2)
+==================
+
+* Fixed issue [#563](https://github.com/jakubpawlowicz/clean-css/issues/563) - `background:inherit` restoring.
+
 [3.2.10 / 2015-05-14](https://github.com/jakubpawlowicz/clean-css/compare/v3.2.9...v3.2.10)
 ==================